pyproject.toml
[project]
name = "faugen-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "deepagents>=0.2.7",
    "dotenv>=0.9.9",
    "embeddings>=0.0.8",
    "faiss-cpu>=1.13.0",
    "fastparquet>=2024.11.0",
    "google-generativeai>=0.8.5",
    "gradio>=5.49.1",
    "langchain>=1.0.7",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.3",
    "lxml>=6.0.2",
    "openai>=2.8.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pip>=25.3",
    "pyarrow>=22.0.0",
    "pymupdf>=1.26.6",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.2",
]
